backend/requirements.txt
# Team members:
# 1. Anusha Gurram – G01514835
# 2. Bala Naga Tirumala Kiran Annadata – G01508483
# 3. Yashwanth Karthikeya Settem– G01512992
# Lists all required Python packages for the FastAPI app, including server, ORM, DB, and configuration tools.

fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlmodel==0.0.14
pydantic-settings==2.1.0
pymysql==1.1.0
alembic==1.13.1
python-dotenv==1.0.1
cryptography==42.0.5
gunicorn==23.0.0